package struct change
