+++ src/main/java/in/zapr/druid/druidry/extensions/histogram/QuantilesPostAggregator.java
@@ -1,9 +1,10 @@
-package in.zapr.druid.druidry.postAggregator;
+package in.zapr.druid.druidry.extensions.histogram;
 
 import java.util.Set;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+import in.zapr.druid.druidry.postAggregator.DruidPostAggregator;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NonNull;
