+++ src/test/java/in/zapr/druid/druidry/extensions/histogram/CustomBucketsPostAggregatorTest.java
@@ -1,4 +1,4 @@
-package in.zapr.druid.druidry.postAggregator;
+package in.zapr.druid.druidry.extensions.histogram;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -16,6 +16,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import in.zapr.druid.druidry.extensions.histogram.CustomBucketsPostAggregator;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
